Reorder Panelizer menu tabs from most used (content) to least used (settings).
This change allows our permissions system to function.  We currently do not
want editors to have access to settings tab but we need them to be able to
change the layout and content.  This fixes an issue where when we have the
above need, the Panelizer tab is not displaying because it defaults to
settings which is the least used and (in our case) least desirable Panelizer
page to land on since our editors most often want to modify the content
placement.